change amout type from long to double to accept floating point

+++ corekit/src/main/java/com/midtrans/sdk/corekit/models/PaymentDetails.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 /**
  * Created by ziahaqi on 2/14/18.
- *
+ * <p>
  * This class provide actual transactions details informations is used by payment pages.
  */
 
@@ -44,7 +44,7 @@ public class PaymentDetails {
         this.itemDetailsList = itemDetailsList;
     }
 
-    public long getTotalAmount() {
+    public double getTotalAmount() {
         return transactionDetails != null ? transactionDetails.getAmount() : 0;
     }
