+++ corekit/src/main/java/com/midtrans/sdk/corekit/models/SnapTransactionDetails.java
@@ -9,18 +9,18 @@ public class SnapTransactionDetails {
     @SerializedName("order_id")
     private String orderId;
     @SerializedName("gross_amount")
-    private int grossAmount;
+    private Double grossAmount;
 
-    public SnapTransactionDetails(String orderId, int grossAmount) {
+    public SnapTransactionDetails(String orderId, Double grossAmount) {
         setOrderId(orderId);
         setGrossAmount(grossAmount);
     }
 
-    public double getGrossAmount() {
+    public Double getGrossAmount() {
         return grossAmount;
     }
 
-    public void setGrossAmount(int grossAmount) {
+    public void setGrossAmount(Double grossAmount) {
         this.grossAmount = grossAmount;
     }
